Refactor Qt compatibility helpers into QtCompat_p.h
Moved Qt version compatibility functions (such as getMetaTypeName, getVariantTypeId, canConvert, and createDefaultVariantForType) from Property.h/cpp and PropertyEditor.cpp into a new internal header QtCompat_p.h. Updated includes throughout the codebase to use the new header, improving maintainability and separation of concerns.
